requirements.txt
affine==2.4.0 ; python_version >= "3.9" and python_version < "4.0"
antlr4-python3-runtime==4.9.3 ; python_version >= "3.9" and python_version < "4.0"
arrow==1.3.0 ; python_version >= "3.9" and python_version < "4.0"
attrs==23.1.0 ; python_version >= "3.9" and python_version < "4.0"
beautifulsoup4==4.12.2 ; python_version >= "3.9" and python_version < "4.0"
binaryornot==0.4.4 ; python_version >= "3.9" and python_version < "4.0"
branca==0.6.0 ; python_version >= "3.9" and python_version < "4.0"
certifi==2023.7.22 ; python_version >= "3.9" and python_version < "4.0"
cftime==1.6.3 ; python_version >= "3.9" and python_version < "4.0"
chardet==5.2.0 ; python_version >= "3.9" and python_version < "4.0"
charset-normalizer==3.3.1 ; python_version >= "3.9" and python_version < "4.0"
click-plugins==1.1.1 ; python_version >= "3.9" and python_version < "4.0"
click==8.1.7 ; python_version >= "3.9" and python_version < "4.0"
cligj==0.7.2 ; python_version >= "3.9" and python_version < "4"
colorama==0.4.6 ; python_version >= "3.9" and python_version < "4.0" and sys_platform == "win32" or python_version >= "3.9" and python_version < "4.0" and platform_system == "Windows"
cookiecutter==2.4.0 ; python_version >= "3.9" and python_version < "4.0"
fiona==1.9.5 ; python_version >= "3.9" and python_version < "4.0"
folium==0.14.0 ; python_version >= "3.9" and python_version < "4.0"
geopandas==0.12.2 ; python_version >= "3.9" and python_version < "4.0"
hydra-core==1.3.2 ; python_version >= "3.9" and python_version < "4.0"
idna==3.4 ; python_version >= "3.9" and python_version < "4.0"
importlib-metadata==6.8.0 ; python_version >= "3.9" and python_version < "3.10"
jinja2==3.1.2 ; python_version >= "3.9" and python_version < "4.0"
libpysal==4.8.0 ; python_version >= "3.9" and python_version < "4.0"
loguru==0.7.2 ; python_version >= "3.9" and python_version < "4.0"
markdown-it-py==3.0.0 ; python_version >= "3.9" and python_version < "4.0"
markupsafe==2.1.3 ; python_version >= "3.9" and python_version < "4.0"
mdurl==0.1.2 ; python_version >= "3.9" and python_version < "4.0"
mesa-geo==0.5.0 ; python_version >= "3.9" and python_version < "4.0"
mesa==1.1.1 ; python_version >= "3.9" and python_version < "4.0"
netcdf4==1.6.4 ; python_version >= "3.9" and python_version < "4.0"
networkx==3.2 ; python_version >= "3.9" and python_version < "4.0"
numpy==1.25.2 ; python_version < "4.0" and python_version >= "3.9"
omegaconf==2.3.0 ; python_version >= "3.9" and python_version < "4.0"
packaging==23.2 ; python_version >= "3.9" and python_version < "4.0"
pandas==1.5.3 ; python_version >= "3.9" and python_version < "4.0"
pendulum==2.1.2 ; python_version >= "3.9" and python_version < "4.0"
pint==0.20.1 ; python_version >= "3.9" and python_version < "4.0"
platformdirs==3.11.0 ; python_version >= "3.9" and python_version < "4.0"
prettytable==3.9.0 ; python_version >= "3.9" and python_version < "4.0"
pygments==2.16.1 ; python_version >= "3.9" and python_version < "4.0"
pyparsing==3.1.1 ; python_version >= "3.9" and python_version < "4.0"
pyproj==3.6.1 ; python_version >= "3.9" and python_version < "4.0"
python-dateutil==2.8.2 ; python_version >= "3.9" and python_version < "4.0"
python-slugify==8.0.1 ; python_version >= "3.9" and python_version < "4.0"
pytz==2023.3.post1 ; python_version >= "3.9" and python_version < "4.0"
pytzdata==2020.1 ; python_version >= "3.9" and python_version < "4.0"
pyyaml==6.0.1 ; python_version >= "3.9" and python_version < "4.0"
rasterio==1.3.9 ; python_version >= "3.9" and python_version < "4.0"
requests==2.31.0 ; python_version >= "3.9" and python_version < "4.0"
rich==13.6.0 ; python_version >= "3.9" and python_version < "4.0"
rioxarray==0.13.4 ; python_version >= "3.9" and python_version < "4.0"
rtree==1.1.0 ; python_version >= "3.9" and python_version < "4.0"
scipy==1.9.3 ; python_version >= "3.9" and python_version < "4.0"
setuptools==68.2.2 ; python_version >= "3.9" and python_version < "4.0"
shapely==2.0.2 ; python_version >= "3.9" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.9" and python_version < "4.0"
snuggs==1.4.7 ; python_version >= "3.9" and python_version < "4.0"
soupsieve==2.5 ; python_version >= "3.9" and python_version < "4.0"
text-unidecode==1.3 ; python_version >= "3.9" and python_version < "4.0"
tornado==6.3.3 ; python_version >= "3.9" and python_version < "4.0"
tqdm==4.66.1 ; python_version >= "3.9" and python_version < "4.0"
types-python-dateutil==2.8.19.14 ; python_version >= "3.9" and python_version < "4.0"
urllib3==2.0.7 ; python_version >= "3.9" and python_version < "4.0"
wcwidth==0.2.8 ; python_version >= "3.9" and python_version < "4.0"
win32-setctime==1.1.0 ; python_version >= "3.9" and python_version < "4.0" and sys_platform == "win32"
xarray==2023.10.1 ; python_version >= "3.9" and python_version < "4.0"
xyzservices==2023.10.0 ; python_version >= "3.9" and python_version < "4.0"
zipp==3.17.0 ; python_version >= "3.9" and python_version < "3.10"
